Assigned users and parts now display on frontend
THe filter list was running one sql query to find tasks, then building the users later on. The getUsers and getParts were returning empty users and parts. To minimize the amount of changes. I designed a new SQL query for filter that incldues the parts and tasks within the results already.

benefits of this change are increased speed (one query to get all results as opposed to one result per query, plus one, and less code to maintain.