Fixed segmentation fault in buildTaskFromJSON.
Replacement functionality is tested to ensure it behaves the same under all circumstances, except for the segmentation fault part. I was unable to run a test longer than 3 seconds without this change. One running GDB, it instantly told me the issue was in the buildTaskFromJSON function. Problem is now solved.